Rank `generic_token` lower than all others.
There is an ambiguity here:
    pipeline ::= expression | command

    expression ::= literal
    literal = integer_literal
    integer_literal = [0..9]+

    command ::= generic_token
    generic_token = Any Unicode character except:
            {		}		(		)		;		,		|		&		$
            `   (The backtick character U+0060)
            double_quote_character
            single_quote_character
            whitespace
            new_line_character

A number (1234) can match either `expression` or `command`. I don't know why this isn't flagged by Irony - beyond
my understanding/memory of parsing theory.

Interestingly, PowerShell allows almost anything to be a command name. For example, this works:
    PS> Set-Alias + Get-Location
    PS> & +

A Terminal has an integer priority. Given the choice between matching `generic_token` and `integer_literal`, I want the latter. In fact,
I want `generic_token` to be a match of last resort. So I set its priorty to -1 (the default is 0).

I couldn't find any discussion of this in the PowerShell Language Specification, but I might have missed it.