Move literal regexes in to productions.
The problem was trying to write matching for integer literals, both hex
and decimal. The pattern described in the spec looks like:

   ([0-9]+)|(0x[0-9a-z]+)

The regex alternation operator `|` is not greedy, so when trying to match
`0xa` it would just match `0`. One fix is to reorder the regex so the hex
part comes before the decimal part, but that's a bit error prone. So
instead I moved the responsibility from the scanner to the parser. I did
that by moving the definitions of the following rules from the regex-based
terminals to the rule-based grammar:

 - `literal`
 - `integer_literal`
 - `string_literal`
 - `real_literal`     # not yet implemented

A consequence of this change is that the AST implementation is now much
simpler. Previously `literal_node.GetValue()` had to figure out which
sub-regex was the best match for this token, but now that's all done by the
parser.

hex integers and expandable strings are still not implemented in the AST;
that's next.