Enumerate arrays on output.
The bug here was that arrays displayed as a type:

    PS> 1..3
    System.Int32[]

This was a consequence of the switch to the new parser. Previously we
would enumerate results for regular commands, but not for the prompt.

Merging these code paths caused the prompt (a string) to print 1 letter
per line (because strings are enumerable). As a temporary fix I disabled
enumerating command output, which caused arrays to print as their type.

Now I changed `PipeLineCommandRuntime.WriteObject` to not enumerating
`string`s, even when you set the `enumerateCollection` flag. So both array
outputs and prompts work again.

Don't enumerate `string` in , ever.

This let's